requirements.txt
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pydevd-pycharm==221.5787.24
Werkzeug==2.1.2
zipp==3.8.0

PyJWT==2.4.0
redis==4.3.3